pyproject.toml
[project]
name = "ai-knowledge-assistant"
version = "0.1.0"
description = "RAG-based AI Personal Knowledge Assistant powered by Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
dependencies = [
    "streamlit>=1.28.0",
    "groq>=0.4.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pypdf2>=3.0.0",
    "python-docx>=1.0.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "gtts>=2.4.0",
    "pydub>=0.25.0",
]

[project.optional-dependencies]
voice = [
    "openai-whisper>=20230918",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
